apps/cli/pyproject.toml
[project]
name = "praxis-cli"
version = "0.1.0"
description = "CLI-first reference client for the Praxis agentic environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.3",
    "litestar>=2.9.0",
    "redis>=5.0.8",
    "pydantic>=2.9.0",
    "structlog>=24.3.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "rich>=13.7.1",
    "uvloop>=0.19.0",
    "httpx>=0.27.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.8"
]
